+++ lib/openhab/dsl/events/watch_event.rb
@@ -9,7 +9,7 @@ module OpenHAB
       # @!attribute [r] path
       #   @return [Pathname] The path that had an event
       # @!attribute [r] type
-      #   @return [:created, :modified, :deleted] Type of change
+      #   @return [:create, :modify, :delete] Type of change
       # @!attribute [r] attachment
       #   @return [Object] The trigger's attachment
       WatchEvent = Struct.new(:type, :path, :attachment)